feat(spa): split admin into Hub and Player destinations (#216)
Phase 2 of #212. The admin SPA now surfaces two distinct top-level
destinations — `/admin/hub` and `/admin/player` — backed by bounded
feature directories that cannot cross-import.

What moved:
- `features/hub-admin/`: Instance, Invitations, Peers, Users, Cache,
  Activity sections (calls existing `/admin/*` endpoints).
- `features/player-admin/`: Sonos + LAN URL section. Gated on a new
  `GET /player/health` probe; when absent, renders a "Player not
  deployed on this host" placeholder and the sidebar hides the link.
- `features/shared/CopyButton`: pure-UI helper used by both sides.

Backend endpoint paths stay on `/admin/*` for this phase — the
`/api/admin/{hub,player}/*` split is #220's job. The boundary
enforcement test (`feature-boundaries.test.ts`) is the tactical guard
until ESLint `no-restricted-paths` lands in #221.
