+++ frontend/src/components/layout/Sidebar.test.tsx
@@ -20,7 +20,16 @@ vi.mock("@/lib/subsonic", async () => {
   };
 });
 
+vi.mock("@/lib/api", async () => {
+  const actual = await vi.importActual<typeof import("@/lib/api")>("@/lib/api");
+  return {
+    ...actual,
+    getPlayerHealth: vi.fn().mockResolvedValue({ status: "ok", appVersion: "test" }),
+  };
+});
+
 import { getMusicFolders } from "@/lib/subsonic";
+import { getPlayerHealth } from "@/lib/api";
 
 function renderSidebar() {
   const qc = new QueryClient({
@@ -38,6 +47,8 @@ function renderSidebar() {
 beforeEach(() => {
   localStorage.clear();
   vi.mocked(getMusicFolders).mockReset();
+  vi.mocked(getPlayerHealth).mockReset();
+  vi.mocked(getPlayerHealth).mockResolvedValue({ status: "ok", appVersion: "test" });
 });
 
 describe("Sidebar Albums group", () => {
@@ -61,6 +72,32 @@ describe("Sidebar Albums group", () => {
 
     const aliceLink = screen.getByText("Alice's Hub").closest("a");
     expect(aliceLink).toHaveAttribute("href", "/library/folder-2");
+  });
+
+  it("surfaces Hub and Player as distinct admin destinations when player-health is present", async () => {
+    vi.mocked(getMusicFolders).mockResolvedValue([]);
+    renderSidebar();
+
+    await waitFor(() => {
+      expect(screen.getByText("Player")).toBeInTheDocument();
+    });
+    expect(screen.getByText("Hub").closest("a")).toHaveAttribute("href", "/admin/hub");
+    expect(screen.getByText("Player").closest("a")).toHaveAttribute("href", "/admin/player");
+  });
+
+  it("hides the Player admin destination when /player/health is absent", async () => {
+    vi.mocked(getMusicFolders).mockResolvedValue([]);
+    vi.mocked(getPlayerHealth).mockResolvedValue(null);
+    renderSidebar();
+
+    await waitFor(() => {
+      expect(screen.getByText("Hub")).toBeInTheDocument();
+    });
+    // Brief settle: the player-health query is async; ensure Player is
+    // never rendered after the gate query resolves.
+    await waitFor(() => {
+      expect(screen.queryByText("Player")).toBeNull();
+    });
   });
 
   it("collapse/expand persists in localStorage", async () => {
