--- frontend/src/pages/AdminPage.tsx
@@ -1,1045 +1,0 @@
-import { useState, useEffect } from "react";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import {
-  getUsers,
-  createUser,
-  deleteUser,
-  getPeers,
-  triggerSync,
-  deletePeerData,
-  generateInvitation,
-  acceptInvitation,
-  getCacheStats,
-  updateCacheSettings,
-  clearArtCache,
-  getInstanceInfo,
-  triggerNavidromeScan,
-  getActivitySettings,
-  updateActivitySettings,
-  getSonosSettings,
-  updateSonosSettings,
-} from "@/lib/api";
-import type { User, Peer } from "@/lib/api";
-import { formatTimeAgo } from "@/lib/format";
-import { cn } from "@/lib/cn";
-import {
-  Plus,
-  Trash2,
-  RefreshCw,
-  Server,
-  Wifi,
-  WifiOff,
-  ImageIcon,
-  Users,
-  Copy,
-  Check,
-  Activity,
-  Speaker,
-} from "lucide-react";
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false);
-
-  const handleCopy = () => {
-    navigator.clipboard.writeText(text).then(() => {
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    });
-  };
-
-  return (
-    <button
-      onClick={handleCopy}
-      title="Copy to clipboard"
-      className="p-1.5 text-text-muted hover:text-text-primary hover:bg-surface-hover rounded transition-colors"
-    >
-      {copied ? <Check className="w-3.5 h-3.5 text-success" /> : <Copy className="w-3.5 h-3.5" />}
-    </button>
-  );
-}
-
-function InstanceSection() {
-  const queryClient = useQueryClient();
-
-  const { data: info, isLoading } = useQuery({
-    queryKey: ["admin-instance"],
-    queryFn: getInstanceInfo,
-    refetchInterval: 30_000,
-  });
-
-  const syncMutation = useMutation({
-    mutationFn: triggerSync,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["admin-instance"] });
-      queryClient.invalidateQueries({ queryKey: ["admin-peers"] });
-      queryClient.invalidateQueries({ queryKey: ["albumList2"] });
-      queryClient.invalidateQueries({ queryKey: ["artists"] });
-    },
-  });
-
-  const scanMutation = useMutation({
-    mutationFn: triggerNavidromeScan,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["admin-instance"] });
-    },
-  });
-
-  if (isLoading || !info) {
-    return (
-      <div className="bg-surface border border-border rounded-lg px-4 py-3">
-        <p className="text-sm text-text-muted">Loading instance info...</p>
-      </div>
-    );
-  }
-
-  const nd = info.navidrome;
-  const isHealthy = nd.reachable && nd.status === "online";
-  const statusLabel = !nd.reachable ? "Unreachable" : nd.status === "online" ? "Online" : nd.status;
-  const statusClass = isHealthy
-    ? "bg-success/10 text-success"
-    : nd.reachable
-    ? "bg-warning/10 text-warning"
-    : "bg-error/10 text-error";
-
-  return (
-    <div className="bg-surface border border-border rounded-lg p-4 space-y-4">
-      {/* Identity — mirrors PeerRow layout */}
-      <div className="px-4 py-3 bg-surface border border-border rounded-lg">
-        <div className="flex items-center gap-4">
-          <Server className="w-5 h-5 text-text-muted shrink-0" />
-          <div className="flex-1 min-w-0">
-            <div className="flex items-center gap-2">
-              <span className="text-sm font-medium text-text-primary truncate">{info.instanceId || "(not set)"}</span>
-              <span className="inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-xs font-medium bg-success/10 text-success">
-                <Wifi className="w-3 h-3" />
-                Online
-              </span>
-            </div>
-            <p className="text-xs text-text-muted truncate">
-              v{info.appVersion} · api {info.apiVersion}
-            </p>
-          </div>
-          {info.instanceId && <CopyButton text={info.instanceId} />}
-        </div>
-        <div className="mt-2 ml-9 flex gap-4 text-xs text-text-secondary">
-          <span>{info.artistCount.toLocaleString()} artists</span>
-          <span>{info.albumCount.toLocaleString()} albums</span>
-          <span>{info.trackCount.toLocaleString()} tracks</span>
-        </div>
-      </div>
-
-      {/* Public Key — copyable */}
-      <div className="space-y-2">
-        <div className="flex items-center gap-2">
-          <Activity className="w-4 h-4 text-text-muted shrink-0" />
-          <span className="text-xs text-text-secondary uppercase tracking-wide font-medium">Public Key</span>
-        </div>
-        <div className="flex items-center gap-2 pl-6">
-          <code className="flex-1 text-xs text-text-primary font-mono bg-surface-hover px-2 py-1.5 rounded break-all leading-relaxed">{info.publicKey}</code>
-          <CopyButton text={info.publicKey} />
-        </div>
-      </div>
-
-      <div className="border-t border-border" />
-
-      {/* Navidrome status */}
-      <div className="space-y-2">
-        <div className="flex items-center justify-between">
-          <span className="text-sm font-medium text-text-primary">Local Navidrome</span>
-          <span className={cn("inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-xs font-medium", statusClass)}>
-            {isHealthy ? <Wifi className="w-3 h-3" /> : <WifiOff className="w-3 h-3" />}
-            {nd.scanning ? "Scanning…" : statusLabel}
-          </span>
-        </div>
-
-        <div className="grid grid-cols-2 gap-3 text-sm">
-          <div>
-            <p className="text-xs text-text-muted mb-0.5">Tracks in index</p>
-            <p className="text-text-primary font-medium">{nd.trackCount.toLocaleString()}</p>
-          </div>
-          <div>
-            <p className="text-xs text-text-muted mb-0.5">Music folders</p>
-            <p className="text-text-primary">{nd.folderCount ?? "—"}</p>
-          </div>
-          <div>
-            <p className="text-xs text-text-muted mb-0.5">Last Navidrome scan</p>
-            <p className="text-text-primary">{formatTimeAgo(nd.lastScan)}</p>
-          </div>
-          <div>
-            <p className="text-xs text-text-muted mb-0.5">Last synced</p>
-            <p className="text-text-primary">{formatTimeAgo(nd.lastSynced)}</p>
-          </div>
-        </div>
-      </div>
-
-      {scanMutation.isSuccess && (
-        <div className="p-3 bg-success/10 border border-success/20 rounded-lg text-sm text-success">
-          Navidrome scan started.
-        </div>
-      )}
-
-      {scanMutation.isError && (
-        <p className="text-sm text-error">
-          {scanMutation.error instanceof Error ? scanMutation.error.message : "Scan failed"}
-        </p>
-      )}
-
-      {syncMutation.isSuccess && (
-        <div className="p-3 bg-success/10 border border-success/20 rounded-lg text-sm text-success">
-          Sync complete — {syncMutation.data.local.trackCount} local tracks indexed.
-        </div>
-      )}
-
-      {syncMutation.isError && (
-        <p className="text-sm text-error">
-          {syncMutation.error instanceof Error ? syncMutation.error.message : "Sync failed"}
-        </p>
-      )}
-
-      <div className="flex gap-2">
-        <button
-          onClick={() => scanMutation.mutate()}
-          disabled={scanMutation.isPending || nd.scanning || !nd.reachable}
-          title="Tell Navidrome to re-scan its music folders"
-          className="flex items-center gap-2 px-3 py-2 bg-surface border border-border hover:bg-surface-hover rounded-lg text-sm text-text-primary transition-colors disabled:opacity-50"
-        >
-          <RefreshCw className={cn("w-4 h-4", (scanMutation.isPending || nd.scanning) && "animate-spin")} />
-          {nd.scanning ? "Scanning…" : "Scan Library"}
-        </button>
-        <button
-          onClick={() => syncMutation.mutate()}
-          disabled={syncMutation.isPending}
-          title="Re-index Navidrome's library into Poutine"
-          className="flex items-center gap-2 px-3 py-2 bg-surface border border-border hover:bg-surface-hover rounded-lg text-sm text-text-primary transition-colors disabled:opacity-50"
-        >
-          <RefreshCw className={cn("w-4 h-4", syncMutation.isPending && "animate-spin")} />
-          {syncMutation.isPending ? "Syncing…" : "Sync Now"}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function PeerRow({ peer }: { peer: Peer }) {
-  const statusConfig =
-    peer.status === "online"
-      ? { className: "bg-success/10 text-success", icon: <Wifi className="w-3 h-3" />, label: "Online" }
-      : { className: "bg-error/10 text-error", icon: <WifiOff className="w-3 h-3" />, label: peer.status };
-
-  const hasCounts = peer.trackCount > 0 || peer.artistCount > 0 || peer.albumCount > 0;
-
-  return (
-    <div className="px-4 py-3 bg-surface border border-border rounded-lg">
-      <div className="flex items-center gap-4">
-        <Server className="w-5 h-5 text-text-muted shrink-0" />
-        <div className="flex-1 min-w-0">
-          <div className="flex items-center gap-2">
-            <span className="text-sm font-medium text-text-primary">{peer.id}</span>
-            <span
-              className={cn(
-                "inline-flex items-center gap-1 px-2 py-0.5 rounded-full text-xs font-medium",
-                statusConfig.className,
-              )}
-            >
-              {statusConfig.icon}
-              {statusConfig.label}
-            </span>
-          </div>
-          <p className="text-xs text-text-muted truncate">{peer.url}</p>
-          {(peer.appVersion || peer.apiVersion) && (
-            <p className="text-xs text-text-muted truncate">
-              {peer.appVersion ? `v${peer.appVersion}` : "unknown version"}
-              {peer.apiVersion !== null ? ` · api ${peer.apiVersion}` : ""}
-            </p>
-          )}
-        </div>
-        <div className="hidden sm:block text-xs text-text-secondary shrink-0">
-          {peer.lastSeen ? `Last seen ${formatTimeAgo(peer.lastSeen)}` : "Never synced"}
-        </div>
-      </div>
-      {hasCounts && (
-        <div className="mt-2 ml-9 flex gap-4 text-xs text-text-secondary">
-          <span>{peer.artistCount.toLocaleString()} artists</span>
-          <span>{peer.albumCount.toLocaleString()} albums</span>
-          <span>{peer.trackCount.toLocaleString()} tracks</span>
-        </div>
-      )}
-      {peer.lastSyncMessage && (
-        <p className={cn(
-          "mt-1.5 ml-9 text-xs",
-          peer.lastSyncOk === false ? "text-error" : "text-text-muted",
-        )}>
-          {peer.lastSyncMessage}
-        </p>
-      )}
-    </div>
-  );
-}
-
-function AddUserForm({ onSuccess }: { onSuccess: () => void }) {
-  const [expanded, setExpanded] = useState(false);
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-
-  const mutation = useMutation({
-    mutationFn: () => createUser(username, password),
-    onSuccess: () => {
-      setUsername("");
-      setPassword("");
-      setExpanded(false);
-      onSuccess();
-    },
-  });
-
-  return (
-    <div className="bg-surface border border-border rounded-lg">
-      <button
-        onClick={() => setExpanded(!expanded)}
-        className="w-full flex items-center gap-2 px-4 py-3 text-sm font-medium text-text-primary hover:bg-surface-hover transition-colors rounded-lg"
-      >
-        <Plus className="w-4 h-4" />
-        Add Guest User
-      </button>
-
-      {expanded && (
-        <form
-          onSubmit={(e) => {
-            e.preventDefault();
-            mutation.mutate();
-          }}
-          className="px-4 pb-4 space-y-3"
-        >
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
-            <div>
-              <label className="block text-sm text-text-secondary mb-1">Username</label>
-              <input
-                type="text"
-                value={username}
-                onChange={(e) => setUsername(e.target.value)}
-                className="w-full px-3 py-2 bg-surface border border-border rounded-lg text-sm text-text-primary focus:outline-none focus:border-accent"
-                required
-              />
-            </div>
-            <div>
-              <label className="block text-sm text-text-secondary mb-1">Password</label>
-              <input
-                type="password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                minLength={8}
-                className="w-full px-3 py-2 bg-surface border border-border rounded-lg text-sm text-text-primary focus:outline-none focus:border-accent"
-                required
-              />
-            </div>
-          </div>
-
-          {mutation.isError && (
-            <p className="text-sm text-error">
-              {mutation.error instanceof Error ? mutation.error.message : "Failed to create user"}
-            </p>
-          )}
-
-          <button
-            type="submit"
-            disabled={mutation.isPending}
-            className="px-4 py-2 bg-accent hover:bg-accent-hover text-white rounded-lg text-sm font-medium transition-colors disabled:opacity-50"
-          >
-            {mutation.isPending ? "Creating..." : "Create User"}
-          </button>
-        </form>
-      )}
-    </div>
-  );
-}
-
-function UserRow({ user, currentUserId }: { user: User; currentUserId: string }) {
-  const queryClient = useQueryClient();
-
-  const deleteMutation = useMutation({
-    mutationFn: () => deleteUser(user.id),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["admin-users"] });
-    },
-  });
-
-  return (
-    <div className="flex items-center gap-4 px-4 py-3 bg-surface border border-border rounded-lg">
-      <Users className="w-5 h-5 text-text-muted shrink-0" />
-      <div className="flex-1 min-w-0">
-        <div className="flex items-center gap-2">
-          <span className="text-sm font-medium text-text-primary">{user.username}</span>
-          {user.isAdmin && (
-            <span className="px-2 py-0.5 bg-accent/10 text-accent rounded-full text-xs font-medium">
-              admin
-            </span>
-          )}
-        </div>
-        <p className="text-xs text-text-muted">Joined {formatTimeAgo(user.createdAt)}</p>
-      </div>
-      {!user.isAdmin && user.id !== currentUserId && (
-        <button
-          onClick={() => {
-            if (window.confirm(`Remove user "${user.username}"?`)) {
-              deleteMutation.mutate();
-            }
-          }}
-          disabled={deleteMutation.isPending}
-          title="Remove user"
-          className="p-2 text-text-muted hover:text-error hover:bg-error/10 rounded-lg transition-colors disabled:opacity-50"
-        >
-          <Trash2 className="w-4 h-4" />
-        </button>
-      )}
-    </div>
-  );
-}
-
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return "0 B";
-  const units = ["B", "KB", "MB", "GB"];
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return `${(bytes / Math.pow(1024, i)).toFixed(i === 0 ? 0 : 1)} ${units[i]}`;
-}
-
-function CacheSection() {
-  const queryClient = useQueryClient();
-  const { data: stats, isLoading } = useQuery({
-    queryKey: ["admin-cache"],
-    queryFn: getCacheStats,
-  });
-
-  const [maxMb, setMaxMb] = useState("");
-  const [dirty, setDirty] = useState(false);
-
-  useEffect(() => {
-    if (stats && !dirty) {
-      setMaxMb(String(Math.round(stats.artCacheMaxBytes / 1024 / 1024)));
-    }
-  }, [stats, dirty]);
-
-  const saveMutation = useMutation({
-    mutationFn: () =>
-      updateCacheSettings({ artCacheMaxBytes: Math.round(parseFloat(maxMb) * 1024 * 1024) }),
-    onSuccess: () => {
-      setDirty(false);
-      queryClient.invalidateQueries({ queryKey: ["admin-cache"] });
-    },
-  });
-
-  const clearMutation = useMutation({
-    mutationFn: clearArtCache,
-    onSuccess: () => queryClient.invalidateQueries({ queryKey: ["admin-cache"] }),
-  });
-
-  if (isLoading || !stats) {
-    return (
-      <div className="bg-surface border border-border rounded-lg px-4 py-3">
-        <p className="text-sm text-text-muted">Loading cache settings...</p>
-      </div>
-    );
-  }
-
-  const usagePercent =
-    stats.artCacheMaxBytes > 0
-      ? Math.min(100, (stats.artCacheCurrentBytes / stats.artCacheMaxBytes) * 100)
-      : 0;
-
-  return (
-    <div className="bg-surface border border-border rounded-lg p-4 space-y-4">
-      <div className="flex items-center gap-2">
-        <ImageIcon className="w-4 h-4 text-text-muted" />
-        <span className="text-sm font-medium text-text-primary">Album Art Cache</span>
-      </div>
-
-      <div>
-        <div className="flex items-center justify-between text-xs text-text-secondary mb-1">
-          <span>
-            {formatBytes(stats.artCacheCurrentBytes)} / {formatBytes(stats.artCacheMaxBytes)}
-          </span>
-          <span>{stats.artCacheFileCount} images</span>
-        </div>
-        <div className="h-2 bg-border rounded-full overflow-hidden">
-          <div
-            className={cn(
-              "h-full rounded-full transition-all",
-              usagePercent > 90 ? "bg-error" : usagePercent > 70 ? "bg-warning" : "bg-accent",
-            )}
-            style={{ width: `${usagePercent}%` }}
-          />
-        </div>
-      </div>
-
-      <div className="flex items-end gap-3">
-        <div className="flex-1">
-          <label className="block text-sm text-text-secondary mb-1">Max Cache Size (MB)</label>
-          <input
-            type="number"
-            min="1"
-            step="1"
-            value={maxMb}
-            onChange={(e) => {
-              setMaxMb(e.target.value);
-              setDirty(true);
-            }}
-            className="w-full px-3 py-2 bg-surface border border-border rounded-lg text-sm text-text-primary focus:outline-none focus:border-accent"
-          />
-        </div>
-        <button
-          onClick={() => saveMutation.mutate()}
-          disabled={!dirty || saveMutation.isPending}
-          className="px-4 py-2 bg-accent hover:bg-accent-hover text-white rounded-lg text-sm font-medium transition-colors disabled:opacity-50"
-        >
-          {saveMutation.isPending ? "Saving..." : "Save"}
-        </button>
-        <button
-          onClick={() => {
-            if (window.confirm("Clear the entire album art cache?")) {
-              clearMutation.mutate();
-            }
-          }}
-          disabled={clearMutation.isPending || stats.artCacheFileCount === 0}
-          className="px-4 py-2 bg-surface border border-border hover:bg-surface-hover rounded-lg text-sm text-text-primary transition-colors disabled:opacity-50"
-        >
-          {clearMutation.isPending ? "Clearing..." : "Clear Cache"}
-        </button>
-      </div>
-
-      {saveMutation.isError && (
-        <p className="text-sm text-error">
-          {saveMutation.error instanceof Error ? saveMutation.error.message : "Failed to save settings"}
-        </p>
-      )}
-    </div>
-  );
-}
-
-function ActivitySettingsSection() {
-  const queryClient = useQueryClient();
-  const { data: settings, isLoading } = useQuery({
-    queryKey: ["admin-activity-settings"],
-    queryFn: getActivitySettings,
-  });
-
-  const [maxEvents, setMaxEvents] = useState("");
-  const [dirty, setDirty] = useState(false);
-
-  useEffect(() => {
-    if (settings && !dirty) {
-      setMaxEvents(String(settings.maxEvents));
-    }
-  }, [settings, dirty]);
-
-  const saveMutation = useMutation({
-    mutationFn: () => updateActivitySettings({ maxEvents: parseInt(maxEvents, 10) }),
-    onSuccess: () => {
-      setDirty(false);
-      queryClient.invalidateQueries({ queryKey: ["admin-activity-settings"] });
-    },
-  });
-
-  if (isLoading || !settings) {
-    return (
-      <div className="bg-surface border border-border rounded-lg px-4 py-3">
-        <p className="text-sm text-text-muted">Loading activity settings...</p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="bg-surface border border-border rounded-lg p-4 space-y-3">
-      <div className="flex items-center gap-2">
-        <Activity className="w-4 h-4 text-text-muted" />
-        <span className="text-sm font-medium text-text-primary">Activity History</span>
-      </div>
-      <p className="text-xs text-text-muted">
-        Maximum number of stored events (streams and syncs each capped independently).
-        Pruned to the most recent N rows when exceeded.
-      </p>
-      <div className="flex items-end gap-3">
-        <div className="flex-1">
-          <label className="block text-sm text-text-secondary mb-1">Max Events</label>
-          <input
-            type="number"
-            min="0"
-            step="1"
-            value={maxEvents}
-            onChange={(e) => {
-              setMaxEvents(e.target.value);
-              setDirty(true);
-            }}
-            className="w-full px-3 py-2 bg-surface border border-border rounded-lg text-sm text-text-primary focus:outline-none focus:border-accent"
-          />
-        </div>
-        <button
-          onClick={() => saveMutation.mutate()}
-          disabled={!dirty || saveMutation.isPending}
-          className="px-4 py-2 bg-accent hover:bg-accent-hover text-white rounded-lg text-sm font-medium transition-colors disabled:opacity-50"
-        >
-          {saveMutation.isPending ? "Saving..." : "Save"}
-        </button>
-      </div>
-      {saveMutation.isError && (
-        <p className="text-sm text-error">
-          {saveMutation.error instanceof Error ? saveMutation.error.message : "Failed to save"}
-        </p>
-      )}
-    </div>
-  );
-}
-
-function SonosSettingsSection() {
-  const queryClient = useQueryClient();
-  const { data: settings, isLoading } = useQuery({
-    queryKey: ["admin-sonos-settings"],
-    queryFn: getSonosSettings,
-  });
-
-  const [volumeCap, setVolumeCap] = useState("");
-  const [dirtyCap, setDirtyCap] = useState(false);
-  const [lanUrl, setLanUrl] = useState("");
-  const [dirtyLan, setDirtyLan] = useState(false);
-
-  useEffect(() => {
-    if (settings && !dirtyCap) {
-      setVolumeCap(String(settings.volumeCap));
-    }
-  }, [settings, dirtyCap]);
-
-  useEffect(() => {
-    if (settings && !dirtyLan) {
-      setLanUrl(settings.lanUrl);
-    }
-  }, [settings, dirtyLan]);
-
-  // Toggle and cap each mutate independently so the toggle UI doesn't get
-  // stuck waiting on a separate "save" press.
-  const toggleMutation = useMutation({
-    mutationFn: (enabled: boolean) => updateSonosSettings({ enabled }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["admin-sonos-settings"] });
-      // Capabilities probe drives DevicePicker visibility — refresh so the
-      // PlayerBar reflects the new state on the next render.
-      queryClient.invalidateQueries({ queryKey: ["capabilities"] });
-    },
-  });
-
-  const capMutation = useMutation({
-    mutationFn: (cap: number) => updateSonosSettings({ volumeCap: cap }),
-    onSuccess: () => {
-      setDirtyCap(false);
-      queryClient.invalidateQueries({ queryKey: ["admin-sonos-settings"] });
-    },
-  });
-
-  const lanMutation = useMutation({
-    mutationFn: (value: string) => updateSonosSettings({ lanUrl: value }),
-    onSuccess: () => {
-      setDirtyLan(false);
-      queryClient.invalidateQueries({ queryKey: ["admin-sonos-settings"] });
-    },
-  });
-
-  if (isLoading || !settings) {
-    return (
-      <div className="bg-surface border border-border rounded-lg px-4 py-3">
-        <p className="text-sm text-text-muted">Loading Sonos settings...</p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="bg-surface border border-border rounded-lg p-4 space-y-3">
-      <div className="flex items-center gap-2">
-        <Speaker className="w-4 h-4 text-text-muted" />
-        <span className="text-sm font-medium text-text-primary">Sonos Casting</span>
-      </div>
-      <p className="text-xs text-text-muted">
-        When enabled, the hub discovers Sonos zones on the LAN and the player
-        gains a device picker. Requires <code>network_mode: host</code> and a
-        LAN URL reachable from Sonos devices (set below — Sonos cannot be
-        enabled until it is). Disabling stops discovery and stops any
-        in-flight casts immediately.
-      </p>
-
-      {/* LAN URL first — every other feature in this section depends on it,
-          so it leads. Sonos enable + DLNA both read from this setting. */}
-      <div className="flex items-end gap-3">
-        <div className="flex-1">
-          <label className="block text-sm text-text-secondary mb-1">
-            LAN URL
-          </label>
-          <input
-            type="url"
-            placeholder="http://192.168.1.10:3000"
-            value={lanUrl}
-            onChange={(e) => {
-              setLanUrl(e.target.value);
-              setDirtyLan(true);
-            }}
-            className="w-full px-3 py-2 bg-surface border border-border rounded-lg text-sm text-text-primary focus:outline-none focus:border-accent"
-          />
-          <p className="mt-1 text-xs text-text-muted">
-            Absolute base URL Sonos + DLNA devices use to fetch streams from
-            this hub. Must be reachable on the LAN (a public hostname or LAN
-            IP, not <code>localhost</code>). Shared with the DLNA MediaServer.
-            Leave empty to disable.
-          </p>
-        </div>
-        <button
-          onClick={() => lanMutation.mutate(lanUrl.trim())}
-          disabled={!dirtyLan || lanMutation.isPending}
-          className="px-4 py-2 bg-accent hover:bg-accent-hover text-white rounded-lg text-sm font-medium transition-colors disabled:opacity-50"
-        >
-          {lanMutation.isPending ? "Saving..." : "Save"}
-        </button>
-      </div>
-
-      <div className="flex items-center justify-between">
-        <div>
-          <p className="text-sm text-text-primary">
-            Status:{" "}
-            <span className={cn(settings.enabled ? "text-success" : "text-text-muted")}>
-              {settings.enabled ? "Enabled" : "Disabled"}
-            </span>
-          </p>
-        </div>
-        <button
-          onClick={() => toggleMutation.mutate(!settings.enabled)}
-          disabled={
-            toggleMutation.isPending ||
-            // Hub rejects enable without a LAN URL (#209) — gate the button
-            // here so the operator sees the dependency, not a 400.
-            (!settings.enabled && !settings.lanUrl)
-          }
-          title={
-            !settings.enabled && !settings.lanUrl
-              ? "Set a LAN URL above before enabling Sonos"
-              : undefined
-          }
-          className={cn(
-            "px-4 py-2 rounded-lg text-sm font-medium transition-colors disabled:opacity-50 disabled:cursor-not-allowed",
-            settings.enabled
-              ? "bg-surface border border-border hover:bg-surface-hover text-text-primary"
-              : "bg-accent hover:bg-accent-hover text-white",
-          )}
-        >
-          {toggleMutation.isPending
-            ? "Saving..."
-            : settings.enabled
-            ? "Disable"
-            : "Enable"}
-        </button>
-      </div>
-
-      <div className="flex items-end gap-3">
-        <div className="flex-1">
-          <label className="block text-sm text-text-secondary mb-1">
-            Volume Cap (0–100)
-          </label>
-          <input
-            type="number"
-            min="0"
-            max="100"
-            step="1"
-            value={volumeCap}
-            onChange={(e) => {
-              setVolumeCap(e.target.value);
-              setDirtyCap(true);
-            }}
-            className="w-full px-3 py-2 bg-surface border border-border rounded-lg text-sm text-text-primary focus:outline-none focus:border-accent"
-          />
-          <p className="mt-1 text-xs text-text-muted">
-            Hard ceiling applied to every SetVolume the hub issues. Sonos
-            volume above this is silently clamped.
-          </p>
-        </div>
-        <button
-          onClick={() => capMutation.mutate(parseInt(volumeCap, 10))}
-          disabled={!dirtyCap || capMutation.isPending}
-          className="px-4 py-2 bg-accent hover:bg-accent-hover text-white rounded-lg text-sm font-medium transition-colors disabled:opacity-50"
-        >
-          {capMutation.isPending ? "Saving..." : "Save"}
-        </button>
-      </div>
-
-      {(toggleMutation.isError || capMutation.isError || lanMutation.isError) && (
-        <p className="text-sm text-error">
-          {(toggleMutation.error || capMutation.error || lanMutation.error) instanceof Error
-            ? (toggleMutation.error || capMutation.error || lanMutation.error)!.message
-            : "Failed to save"}
-        </p>
-      )}
-    </div>
-  );
-}
-
-
-/**
- * v5 invitation flow: generate a signed invite, or paste one received
- * out-of-band to admit the inviter as a peer. Discovered peers (via gossip)
- * appear in the peers list automatically after the next sync round.
- */
-function InvitationsSection({ onPeerAdded }: { onPeerAdded: () => void }) {
-  const [ourUrl, setOurUrl] = useState<string>(
-    typeof window !== "undefined" ? window.location.origin : "",
-  );
-  const [inviteeUrl, setInviteeUrl] = useState("");
-  const [generated, setGenerated] = useState<string>("");
-  const [pasted, setPasted] = useState("");
-  const [acceptResult, setAcceptResult] = useState<string>("");
-
-  const generate = useMutation({
-    mutationFn: () =>
-      generateInvitation({
-        ourUrl,
-        inviteeUrl: inviteeUrl || undefined,
-      }),
-    onSuccess: (data) => setGenerated(data.invitation),
-  });
-
-  const accept = useMutation({
-    mutationFn: () => acceptInvitation({ invitation: pasted.trim(), ourUrl }),
-    onSuccess: (data) => {
-      setAcceptResult(`Admitted peer: ${data.peerId} (${data.peerUrl})`);
-      setPasted("");
-      onPeerAdded();
-    },
-    onError: (err) =>
-      setAcceptResult(err instanceof Error ? err.message : String(err)),
-  });
-
-  return (
-    <div className="space-y-4 mb-6 p-4 rounded-lg border border-border bg-surface">
-      <div>
-        <label className="block text-sm text-text-muted mb-1">
-          This hub's public URL
-        </label>
-        <input
-          type="text"
-          value={ourUrl}
-          onChange={(e) => setOurUrl(e.target.value)}
-          placeholder="https://my-hub.example"
-          className="w-full px-3 py-2 bg-background border border-border rounded text-sm text-text-primary"
-        />
-        <p className="mt-1 text-xs text-text-muted">
-          Used in both flows: when generating an invite it's embedded so the
-          invitee knows where to handshake back; when accepting one it's the
-          URL we send to the inviter so they know how to reach us.
-          Defaults to this browser's origin — override if peers reach this
-          hub at a different address.
-        </p>
-      </div>
-
-      <div className="grid md:grid-cols-2 gap-4">
-        <div className="space-y-2">
-          <h3 className="text-sm font-semibold text-text-primary">Generate invite</h3>
-          <input
-            type="text"
-            value={inviteeUrl}
-            onChange={(e) => setInviteeUrl(e.target.value)}
-            placeholder="Invitee URL (optional — leave blank for open invite)"
-            className="w-full px-3 py-2 bg-background border border-border rounded text-sm text-text-primary"
-          />
-          <button
-            onClick={() => generate.mutate()}
-            disabled={generate.isPending || !ourUrl}
-            className="px-3 py-2 bg-surface-hover border border-border rounded text-sm text-text-primary disabled:opacity-50"
-          >
-            {generate.isPending ? "Signing..." : "Generate"}
-          </button>
-          {generated && (
-            <div className="relative">
-              <textarea
-                readOnly
-                value={generated}
-                onClick={(e) => (e.target as HTMLTextAreaElement).select()}
-                rows={4}
-                className="w-full px-3 py-2 pr-10 bg-background border border-border rounded text-xs font-mono text-text-primary"
-              />
-              <div className="absolute top-1.5 right-1.5">
-                <CopyButton text={generated} />
-              </div>
-            </div>
-          )}
-        </div>
-
-        <div className="space-y-2">
-          <h3 className="text-sm font-semibold text-text-primary">Accept invite</h3>
-          <textarea
-            value={pasted}
-            onChange={(e) => setPasted(e.target.value)}
-            placeholder="Paste invitation here"
-            rows={4}
-            className="w-full px-3 py-2 bg-background border border-border rounded text-xs font-mono text-text-primary"
-          />
-          <button
-            onClick={() => accept.mutate()}
-            disabled={accept.isPending || !pasted.trim() || !ourUrl}
-            className="px-3 py-2 bg-surface-hover border border-border rounded text-sm text-text-primary disabled:opacity-50"
-          >
-            {accept.isPending ? "Verifying..." : "Accept"}
-          </button>
-          {acceptResult && (
-            <p
-              className={cn(
-                "text-xs",
-                accept.isError ? "text-error" : "text-success",
-              )}
-            >
-              {acceptResult}
-            </p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export function AdminPage() {
-  const queryClient = useQueryClient();
-
-  const { data: peers, isLoading: peersLoading } = useQuery({
-    queryKey: ["admin-peers"],
-    queryFn: getPeers,
-  });
-
-  const { data: users, isLoading: usersLoading } = useQuery({
-    queryKey: ["admin-users"],
-    queryFn: getUsers,
-  });
-
-  const syncMutation = useMutation({
-    mutationFn: triggerSync,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["admin-peers"] });
-      queryClient.invalidateQueries({ queryKey: ["admin-instance"] });
-      queryClient.invalidateQueries({ queryKey: ["albumList2"] });
-      queryClient.invalidateQueries({ queryKey: ["artists"] });
-    },
-  });
-
-  const deletePeerDataMutation = useMutation({
-    mutationFn: deletePeerData,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["admin-peers"] });
-    },
-  });
-
-  const currentUserId = users?.find((u) => u.isAdmin)?.id ?? "";
-
-  return (
-    <div className="max-w-4xl mx-auto px-4 py-8 space-y-10">
-      {/* Instance */}
-      <section>
-        <h1 className="text-xl font-bold text-text-primary mb-4">This Instance</h1>
-        <InstanceSection />
-      </section>
-
-      {/* Invitations — generate or accept (above Federation Peers) */}
-      <section>
-        <h1 className="text-xl font-bold text-text-primary mb-4">Invitations</h1>
-        <InvitationsSection
-          onPeerAdded={() => queryClient.invalidateQueries({ queryKey: ["admin-peers"] })}
-        />
-      </section>
-
-      {/* Peers */}
-      <section>
-        <div className="flex items-center justify-between mb-4">
-          <h1 className="text-xl font-bold text-text-primary">Federation Peers</h1>
-          <div className="flex items-center gap-2">
-            <button
-              onClick={() => syncMutation.mutate()}
-              disabled={syncMutation.isPending}
-              className="flex items-center gap-2 px-3 py-2 bg-surface border border-border hover:bg-surface-hover rounded-lg text-sm text-text-primary transition-colors disabled:opacity-50"
-            >
-              <RefreshCw className={cn("w-4 h-4", syncMutation.isPending && "animate-spin")} />
-              {syncMutation.isPending ? "Syncing..." : "Sync All"}
-            </button>
-            {peers && peers.length > 0 && (
-              <button
-                onClick={() => {
-                  if (window.confirm("Delete all data fetched from peers? This will reset sync state.")) {
-                    deletePeerDataMutation.mutate();
-                  }
-                }}
-                disabled={deletePeerDataMutation.isPending}
-                className="flex items-center gap-2 px-3 py-2 bg-surface border border-error/40 hover:bg-error/10 rounded-lg text-sm text-error transition-colors disabled:opacity-50"
-              >
-                <Trash2 className="w-4 h-4" />
-                {deletePeerDataMutation.isPending ? "Deleting..." : "Delete Peer Data"}
-              </button>
-            )}
-          </div>
-        </div>
-
-        {syncMutation.isSuccess && (() => {
-          const failedPeers = syncMutation.data.peers.filter((p) => p.errors.length > 0);
-          if (failedPeers.length > 0) {
-            return (
-              <div className="mb-4 p-3 bg-error/10 border border-error/20 rounded-lg text-sm text-error space-y-1">
-                <p>Sync finished with errors — {failedPeers.length} peer(s) failed:</p>
-                {failedPeers.map((p) => (
-                  <p key={p.instanceId} className="ml-2">{p.instanceId}: {p.errors[0]}</p>
-                ))}
-              </div>
-            );
-          }
-          return (
-            <div className="mb-4 p-3 bg-success/10 border border-success/20 rounded-lg text-sm text-success">
-              Sync complete — local: {syncMutation.data.local.trackCount} tracks,{" "}
-              {syncMutation.data.peers.length} peer(s) synced.
-            </div>
-          );
-        })()}
-
-        {syncMutation.isError && (
-          <p className="mb-4 text-sm text-error">
-            {syncMutation.error instanceof Error ? syncMutation.error.message : "Sync failed"}
-          </p>
-        )}
-
-        <div className="space-y-2">
-          {peersLoading && <p className="text-sm text-text-muted py-4">Loading peers...</p>}
-          {!peersLoading && peers?.length === 0 && (
-            <p className="text-sm text-text-muted py-4">
-              No peers yet. Generate an invite above and send it to the other operator,
-              or paste one you've received.
-            </p>
-          )}
-          {peers?.map((peer) => <PeerRow key={peer.id} peer={peer} />)}
-        </div>
-      </section>
-
-      {/* Users */}
-      <section>
-        <h2 className="text-xl font-bold text-text-primary mb-4">Users</h2>
-        <div className="space-y-2">
-          <AddUserForm onSuccess={() => queryClient.invalidateQueries({ queryKey: ["admin-users"] })} />
-          {usersLoading && <p className="text-sm text-text-muted py-4">Loading users...</p>}
-          {users?.map((user) => (
-            <UserRow key={user.id} user={user} currentUserId={currentUserId} />
-          ))}
-        </div>
-      </section>
-
-      {/* Cache */}
-      <section>
-        <h2 className="text-xl font-bold text-text-primary mb-4">Cache</h2>
-        <CacheSection />
-      </section>
-
-      {/* Activity history retention */}
-      <section>
-        <h2 className="text-xl font-bold text-text-primary mb-4">Activity</h2>
-        <ActivitySettingsSection />
-      </section>
-
-      {/* Sonos casting (#184) */}
-      <section>
-        <h2 className="text-xl font-bold text-text-primary mb-4">Sonos</h2>
-        <SonosSettingsSection />
-      </section>
-    </div>
-  );
-}
